add Error500 component to handle server failures (i.e. instances of netlify functions failing to communicate with airtable successfully) add serverError state to Report to direct users back to Splitter remove fallback data for questions and dividers from Report
