--- src/App/Loading/style.js
@@ -1,14 +1,0 @@
-import styled from 'styled-components';
-
-const CenterWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 100vw;
-  height: 100vh;
-  margin: auto;
-  align-items: center;
-  justify-content: center;
-  align-content: center;
-`;
-
-export { CenterWrapper };
