+++ src/App/Report/Report.js
@@ -9,19 +9,11 @@ import { useLocation } from 'react-router-dom';
 
 // import subcomponents and reusables
 import { Form, Divider, Confirm, Submit } from './index';
-import { Loading } from '../index';
+import { Loading, Error500 } from '../index';
 
 // and utils and libraries
 import { getData, deleteValue, findIndex } from '../../utils';
 import uuid from 'uuid/v4';
-
-// and fallback data
-import {
-  firstQuestions,
-  witnessQuestions,
-  firstDividers,
-  witnessDividers,
-} from '../../model';
 
 const Report = () => {
   // grab React Router state to determine which components to render at Report level, and which questions/dividers to fetch
@@ -37,6 +29,7 @@ const Report = () => {
   // set up states
   const [questions, setQuestions] = useState(null);
   const [dividers, setDividers] = useState(null);
+  const [serverError, setServerError] = useState(false);
 
   // generate uuid as userRef and memoize (to be passed into Submit and Confirm)
   const userRef = useMemo(() => uuid(), []);
@@ -47,11 +40,7 @@ const Report = () => {
         setQuestions(records);
       })
       .catch((err) => {
-        setQuestions(choice === 'first' ? firstQuestions : witnessQuestions);
-        // console.log(
-        //   'Failed to fetch question data - falling back to hard coding. Error: ',
-        //   err
-        // );
+        setServerError(true);
       });
 
     getData(`${choice}-dividers`)
@@ -59,11 +48,7 @@ const Report = () => {
         setDividers(records);
       })
       .catch((err) => {
-        setDividers(choice === 'first' ? firstDividers : witnessDividers);
-        // console.log(
-        //   'Failed to fetch divider data - falling back to hard coding. Error: ',
-        //   err
-        // );
+        setServerError(true);
       });
     // eslint-disable-next-line
   }, []); // simiarly if we write [choice] here, we'll repeatedly fetch the same data
@@ -151,7 +136,13 @@ const Report = () => {
     });
   }, []);
 
-  // if any API calls have yet to resolve, render Loading component
+  // if either API call explicitly fails, render the 500 error and prompt user to reload
+  // this will only happen, as if they proceed past this point the data must be available
+  if (serverError) {
+    return <Error500 pathname='/choose' />;
+  }
+
+  // if the API calls have simply yet to resolve, render Loading component
   if (!(questions && dividers)) return <Loading />;
 
   if (location.pathname.includes('section')) {
